querier/requirements.txt
requests~=2.32.3
dotenv~=0.9.9
python-dotenv~=1.1.0
tabulate~=0.9.0
pip~=25.0.1
idna~=3.10
certifi~=2025.4.26
charset-normalizer~=3.4.2
urllib3~=2.4.0
pytz~=2025.2
pynacl~=1.5.0